+++ src/main/java/edu/wpi/GoldenGandaberundas/tableControllers/AudioVisualService/AudioVisualClientServer.java
@@ -14,13 +14,13 @@ import java.util.stream.Collectors;
 
 public class AudioVisualClientServer implements TableController<AudioVisual, Integer> {
   /** name of table */
-  protected String tbName;
+  private String tbName;
   /** name of columns in database table the first entry is the primary key */
-  protected List<String> colNames;
+  private List<String> colNames;
   /** list of keys that make a composite primary key */
-  protected String pkCols = null;
+  private String pkCols = null;
   /** list that contains the objects stored in the database */
-  protected ArrayList<AudioVisual> objList;
+  private ArrayList<AudioVisual> objList;
   /** relative path to the database file */
   ConnectionHandler connectionHandler = ConnectionHandler.getInstance();
