+++ src/main/java/edu/wpi/GoldenGandaberundas/tableControllers/GiftDeliveryService/GiftClientServer.java
@@ -13,13 +13,13 @@ import java.util.Locale;
 import java.util.stream.Collectors;
 
 public class GiftClientServer implements TableController<Gift, Integer> {
-  protected String tbName;
+  private String tbName;
   /** name of columns in database table the first entry is the primary key */
-  protected List<String> colNames;
+  private List<String> colNames;
   /** list of keys that make a composite primary key */
-  protected String pkCols = null;
+  private String pkCols = null;
   /** list that contains the objects stored in the database */
-  protected ArrayList<Gift> objList;
+  private ArrayList<Gift> objList;
   /** relative path to the database file */
   ConnectionHandler connectionHandler = ConnectionHandler.getInstance();
