+++ src/main/java/edu/wpi/GoldenGandaberundas/tableControllers/GiftDeliveryService/GiftRequestClientServer.java
@@ -1,0 +1,4 @@
+package edu.wpi.GoldenGandaberundas.tableControllers.GiftDeliveryService;
+
+public class GiftRequestClientServer {
+}
