+++ src/main/java/edu/wpi/GoldenGandaberundas/tableControllers/GiftDeliveryService/GiftRequestEmbedded.java
@@ -1,0 +1,4 @@
+package edu.wpi.GoldenGandaberundas.tableControllers.GiftDeliveryService;
+
+public class GiftRequestEmbedded {
+}
