+++ src/main/java/edu/wpi/GoldenGandaberundas/tableControllers/LaundryService/LaundryRequestTbl.java
@@ -2,6 +2,7 @@ package edu.wpi.GoldenGandaberundas.tableControllers.LaundryService;
 
 import edu.wpi.GoldenGandaberundas.TableController;
 import edu.wpi.GoldenGandaberundas.tableControllers.DBConnection.ConnectionHandler;
+import edu.wpi.GoldenGandaberundas.tableControllers.DBConnection.ConnectionType;
 import edu.wpi.GoldenGandaberundas.tableControllers.Requests.Request;
 import edu.wpi.GoldenGandaberundas.tableControllers.Requests.RequestTable;
 import java.io.*;
@@ -25,6 +26,10 @@ public class LaundryRequestTbl implements TableController<LaundryRequest, ArrayL
   /** list that contains the objects stored in the database */
   protected ArrayList<LaundryRequest> objList;
   /** relative path to the database file */
+  TableController<LaundryRequest, ArrayList<Integer>> embeddedTable = null;
+
+  TableController<LaundryRequest, ArrayList<Integer>> clientServerTable = null;
+
   ConnectionHandler connectionHandler = ConnectionHandler.getInstance();
 
   Connection connection = connectionHandler.getConnection();
@@ -34,9 +39,15 @@ public class LaundryRequestTbl implements TableController<LaundryRequest, ArrayL
     pkCols = "reqID, laundryID";
     String[] cols = {"reqID", "laundryID"};
     colNames = Arrays.asList(cols);
+    objList = new ArrayList<LaundryRequest>();
+    embeddedTable =
+            new LaundryRequestEmbedded(tbName, colNames.toArray(new String[4]), pkCols, objList);
+    clientServerTable =
+            new LaundryRequestClientServer(tbName, colNames.toArray(new String[4]), pkCols, objList);
+    connectionHandler.addTable(embeddedTable, ConnectionType.embedded);
+    connectionHandler.addTable(clientServerTable, ConnectionType.clientServer);
     masterTable = RequestTable.getInstance();
     createTable();
-    objList = new ArrayList<LaundryRequest>();
     objList = readTable();
   }
 
@@ -56,234 +67,41 @@ public class LaundryRequestTbl implements TableController<LaundryRequest, ArrayL
     return instance;
   }
 
-  @Override
-  public ArrayList<LaundryRequest> readTable() {
-    ArrayList tableInfo = new ArrayList<LaundryRequest>(); // **
-    try {
-      PreparedStatement s = connection.prepareStatement("SElECT * FROM " + tbName + ";");
-      ResultSet r = s.executeQuery();
-      while (r.next()) {
-        System.out.println(masterTable.getEntry(r.getInt(1)));
-        tableInfo.add(
-            new LaundryRequest( // **
-                masterTable.getEntry(r.getInt(1)), r.getInt(2)));
-      }
-    } catch (SQLException se) {
-      se.printStackTrace();
-      return null;
+  private TableController<LaundryRequest, ArrayList<Integer>> getCurrentTable() {
+    System.out.println("Connection Type: " + connectionHandler.getCurrentConnectionType());
+    switch (connectionHandler.getCurrentConnectionType()) {
+      case embedded:
+        return embeddedTable;
+      case clientServer:
+        return clientServerTable;
+      case cloud:
+        return null;
     }
-    return tableInfo;
+    System.out.println(connectionHandler.getCurrentConnectionType());
+    return null;
   }
 
   @Override
-  public boolean addEntry(LaundryRequest obj) {
-    if (!RequestTable.getInstance().entryExists(obj.getRequestID())) {
-      RequestTable.getInstance().addEntry(obj);
-    }
-    LaundryRequest launReq = (LaundryRequest) obj; // **
-    PreparedStatement s = null;
-    try {
-      s =
-          connection.prepareStatement( // **
-              "INSERT OR IGNORE INTO " + tbName + " VALUES (?, ?);");
-
-      // **
-      s.setInt(1, launReq.getRequestID());
-      s.setInt(2, launReq.getLaundryID());
-      s.executeUpdate();
-      return true;
-    } catch (SQLException e) {
-      e.printStackTrace();
-      return false;
-    }
-  }
-
-  private boolean addEntryOnline(LaundryRequest laundryReq) {
-    try {
-      PreparedStatement s =
-          connection.prepareStatement(
-              " IF NOT EXISTS (SELECT 1 FROM "
-                  + tbName
-                  + " WHERE "
-                  + colNames.get(0)
-                  + " = ?"
-                  + " AND "
-                  + colNames.get(1)
-                  + " = ?"
-                  + ")"
-                  + "BEGIN"
-                  + "    INSERT INTO "
-                  + tbName
-                  + " VALUES (?, ?)"
-                  + "end");
-
-      s.setInt(1, laundryReq.getRequestID());
-      s.setInt(2, laundryReq.getLaundryID());
-      s.setInt(3, laundryReq.getRequestID());
-      s.setInt(4, laundryReq.getLaundryID());
-      s.executeUpdate();
-      return true;
-    } catch (SQLException e) {
-      e.printStackTrace();
-      return false;
-    }
-  }
+  public ArrayList<LaundryRequest> readTable() {return this.getCurrentTable().readTable();}
 
   @Override
-  public ArrayList<LaundryRequest> readBackup(String fileName) {
-    ArrayList<LaundryRequest> laundryList = new ArrayList<LaundryRequest>(); // **
-
-    try {
-      File csvFile = new File(fileName);
-      BufferedReader buffer = new BufferedReader(new FileReader(csvFile)); // reads the files
-      String currentLine = buffer.readLine(); // reads a line from the csv file
-      System.out.println(currentLine);
-
-      currentLine = buffer.readLine();
-
-      while (currentLine != null) { // cycles in the while loop until it reaches the end
-        String[] element = currentLine.split(","); // separates each element based on a comma
-        LaundryRequest req = // **
-            new LaundryRequest(
-                Integer.parseInt(element[0]),
-                element[1],
-                Integer.parseInt(element[2]),
-                Integer.parseInt(element[3]),
-                Long.parseLong(element[4]),
-                Long.parseLong(element[5]),
-                element[8],
-                element[9],
-                Integer.parseInt(element[10])); // **
-        laundryList.add(req);
-        currentLine = buffer.readLine();
-      }
-      ; // creates a Location
-
-    } catch (FileNotFoundException ex) {
-      ex.printStackTrace();
-    } catch (IOException ex) {
-      ex.printStackTrace();
-    }
-    return laundryList; // **
-  }
+  public boolean addEntry(LaundryRequest obj) {return this.getCurrentTable().addEntry(obj);}
 
   @Override
-  public void createTable() {
-    try {
-      PreparedStatement s =
-          connection.prepareStatement(
-              "SELECT count(*) FROM sqlite_master WHERE tbl_name = ? LIMIT 1;");
-      s.setString(1, tbName);
-      ResultSet r = s.executeQuery();
-      r.next();
-      if (r.getInt(1) != 0) {
-        return;
-      }
-    } catch (SQLException e) {
-      e.printStackTrace();
-      return;
-    }
-
-    try {
-      Class.forName("org.sqlite.JDBC");
-    } catch (ClassNotFoundException e) {
-      System.out.println("SQLite driver not found on classpath, check your gradle configuration.");
-      e.printStackTrace();
-      return;
-    }
-
-    System.out.println("SQLite driver registered!");
-
-    Statement s = null;
-    try {
-      s = connection.createStatement();
-      s.execute("PRAGMA foreign_keys = ON"); // **
-      s.execute(
-          "CREATE TABLE IF NOT EXISTS  LaundryRequests("
-              + "reqID INTEGER NOT NULL, "
-              + "laundryID INTEGER NOT NULL, "
-              + "CONSTRAINT LaundryRequestPK PRIMARY KEY (reqID, laundryID), "
-              + "CONSTRAINT LaundryRequestFK1 FOREIGN KEY (reqID) REFERENCES Requests (requestID) "
-              + "ON UPDATE CASCADE "
-              + "ON DELETE CASCADE , "
-              + "CONSTRAINT LaundryRequestFK2 FOREIGN KEY (laundryID) REFERENCES Laundry (laundryID) "
-              + "ON UPDATE CASCADE "
-              + "ON DELETE CASCADE"
-              + ");");
-
-    } catch (SQLException e) {
-      e.printStackTrace();
-    }
-  }
-
-  private void createTableOnline() {
-    try {
-
-      PreparedStatement s1 =
-          connection.prepareStatement("SELECT COUNT(*) FROM sys.tables WHERE name = ?;");
-      s1.setString(1, tbName);
-      ResultSet r = s1.executeQuery();
-      r.next();
-      if (r.getInt(1) != 0) {
-        return;
-      }
-      Statement s = connection.createStatement();
-      s.execute(
-          "CREATE TABLE  LaundryRequests("
-              + "reqID INTEGER NOT NULL, "
-              + "laundryID INTEGER NOT NULL, "
-              + "CONSTRAINT LaundryRequestPK PRIMARY KEY (reqID, laundryID), "
-              + "CONSTRAINT LaundryRequestFK1 FOREIGN KEY (reqID) REFERENCES Requests (requestID) "
-              + "ON UPDATE CASCADE "
-              + "ON DELETE CASCADE , "
-              + "CONSTRAINT LaundryRequestFK2 FOREIGN KEY (laundryID) REFERENCES Laundry (laundryID) "
-              + "ON UPDATE CASCADE "
-              + "ON DELETE CASCADE"
-              + ");");
-    } catch (SQLException e) {
-      e.printStackTrace();
-    }
-  }
+  public ArrayList<LaundryRequest> readBackup(String fileName) {return this.getCurrentTable().readBackup(fileName);}
 
   @Override
-  public LaundryRequest getEntry(ArrayList<Integer> pkID) {
-    LaundryRequest launReq = new LaundryRequest();
+  public void createTable() {this.getCurrentTable().createTable();}
 
-    if (this.entryExists(pkID)) {
-      try {
-        PreparedStatement s =
-            connection.prepareStatement(
-                "SELECT * FROM " + tbName + " WHERE (" + pkCols + ") =(?,?);");
-        s.setInt(1, pkID.get(0));
-        s.setInt(2, pkID.get(1));
-        ResultSet r = s.executeQuery();
-        r.next(); // **
-
-        if (entryExists(pkID)) {
-          launReq.setRequestID(r.getInt(1));
-          launReq.setLaundryID(r.getInt(2));
-        }
-
-        return launReq; // **
-      } catch (SQLException e) {
-        e.printStackTrace();
-      }
-    }
-    return launReq; // **
-  }
+  @Override
+  public LaundryRequest getEntry(ArrayList<Integer> pkID) {return this.getCurrentTable().getEntry(pkID);}
 
   @Override
   public boolean loadFromArrayList(ArrayList<LaundryRequest> objList) {
-    return false;
+    return this.getCurrentTable().loadFromArrayList(objList);
   }
 
-  public void writeTable() {
-
-    for (LaundryRequest obj : objList) {
-
-      this.addEntry(obj);
-    }
-  }
+  public void writeTable() {this.getCurrentTable().writeTable();}
 
   /**
    * Modifies the attribute so that it is equal to value MAKE SURE YOU KNOW WHAT DATA TYPE YOU ARE
@@ -295,32 +113,8 @@ public class LaundryRequestTbl implements TableController<LaundryRequest, ArrayL
    * @return true if successful, false otherwise
    */
   // public boolean editEntry(T1 pkid, String colName, Object value)
-  public boolean editEntry(ArrayList<Integer> pkid, String colName, Object value) {
-    StringBuilder pkString = new StringBuilder();
-    for (int i = 0; i < pkid.size() - 1; i++) {
-      pkString.append(pkid.get(i)).append(",");
-    }
-    pkString.append(pkid.get(pkid.size() - 1));
-    try {
-      PreparedStatement s =
-          connection.prepareStatement(
-              "UPDATE "
-                  + tbName
-                  + " SET "
-                  + colName
-                  + " = ? WHERE ("
-                  + pkCols
-                  + ") = ("
-                  + pkString
-                  + ");");
-      s.setObject(1, value);
-      s.executeUpdate();
-      return true;
-    } catch (SQLException e) {
-      e.printStackTrace();
-      return false;
-    }
-  }
+  public boolean editEntry(ArrayList<Integer> pkid, String colName, Object value) {return this.getCurrentTable().editEntry(pkid, colName, value);}
+
 
   /**
    * removes a row from the database
@@ -328,125 +122,20 @@ public class LaundryRequestTbl implements TableController<LaundryRequest, ArrayL
    * @param pkid primary key of row to be removed
    * @return true if successful, false otherwise
    */
-  public boolean deleteEntry(ArrayList<Integer> pkid) {
-    StringBuilder pkString = new StringBuilder();
-    for (int i = 0; i < pkid.size() - 1; i++) {
-      pkString.append(pkid.get(i)).append(",");
-    }
-    pkString.append(pkid.get(pkid.size() - 1));
-    try {
-      PreparedStatement s =
-          connection.prepareStatement(
-              "DELETE FROM " + tbName + " WHERE (" + pkCols + ") = (" + pkString + ");");
-      s.executeUpdate();
-      return true;
-    } catch (SQLException e) {
-      e.printStackTrace();
-    }
-    return false;
-  }
+  public boolean deleteEntry(ArrayList<Integer> pkid) {return this.getCurrentTable().deleteEntry(pkid);}
 
   /**
    * creates CSV file representing the objects stored in the table
    *
    * @param f filename of the to be created CSV
    */
-  public void createBackup(File f) {
-    if (objList.isEmpty()) {
-      return;
-    }
-    /* Instantiate the writer */
-    PrintWriter writer = null;
-    try {
-      writer = new PrintWriter(f);
-    } catch (FileNotFoundException e) {
-      e.printStackTrace();
-    }
-
-    /* Get the class type of the objects in the array */
-    final Class<?> type = objList.get(0).getClass();
-
-    /* Get the name of all the attributes */
-    final ArrayList<Field> classAttributes = new ArrayList<>(List.of(type.getDeclaredFields()));
-
-    boolean doesExtend = Request.class.isAssignableFrom(type);
-    if (doesExtend) {
-      final Class<?> superType = objList.get(0).getClass().getSuperclass();
-      classAttributes.addAll(0, (List.of(superType.getDeclaredFields())));
-    }
-
-    /* Write the parsed attributes to the file */
-    writer.println(classAttributes.stream().map(Field::getName).collect(Collectors.joining(",")));
-
-    /* For each object, read each attribute and append it to the file with a comma separating */
-    PrintWriter finalWriter = writer;
-    objList.forEach(
-        obj -> {
-          finalWriter.println(
-              classAttributes.stream()
-                  .map(
-                      attribute -> {
-                        attribute.setAccessible(true);
-                        String output = "";
-                        try {
-                          output = attribute.get(obj).toString();
-                        } catch (IllegalAccessException | ClassCastException e) {
-                          System.err.println("[CSVUtil] Object attribute access error.");
-                        }
-                        return output;
-                      })
-                  .collect(Collectors.joining(",")));
-          finalWriter.flush();
-        });
-    writer.close();
-  }
+  public void createBackup(File f) {this.getCurrentTable().createBackup(f);}
 
   // drop current table and enter data from CSV
-  public ArrayList<LaundryRequest> loadBackup(String fileName) {
-    createTable();
-    ArrayList<LaundryRequest> listObjs = readBackup(fileName);
-
-    try {
-      PreparedStatement s = connection.prepareStatement("DELETE FROM " + tbName + ";");
-      s.executeUpdate();
-      this.objList = listObjs;
-      this.writeTable();
-    } catch (SQLException e) {
-      e.printStackTrace();
-    }
-    return listObjs;
-  }
+  public ArrayList<LaundryRequest> loadBackup(String fileName) {return this.getCurrentTable().loadBackup(fileName);}
 
   // checks if an entry exists
-  public boolean entryExists(ArrayList<Integer> pkID) {
-    boolean exists = false;
-    StringBuilder pkString = new StringBuilder();
-    for (int i = 0; i < pkID.size() - 1; i++) {
-      pkString.append(pkID.get(i)).append(",");
-    }
-    pkString.append(pkID.get(pkID.size() - 1));
-
-    try {
-      PreparedStatement s =
-          connection.prepareStatement(
-              "SELECT count(*) FROM "
-                  + tbName
-                  + " WHERE ("
-                  + pkCols
-                  + ") = ("
-                  + pkString.toString()
-                  + ");");
-      ResultSet r = s.executeQuery();
-      r.next();
-      if (r.getInt(1) != 0) {
-        exists = true;
-      }
-      return exists;
-    } catch (SQLException e) {
-      e.printStackTrace();
-      return false;
-    }
-  }
+  public boolean entryExists(ArrayList<Integer> pkID) {return this.getCurrentTable().entryExists(pkID);}
 
   public String getTableName() {
     return tbName;
